@@ -1,9 +1,9 @@
 // @flow
 
-import { logError } from '../utils/common'
+import { logError, invariant } from '../../utils/common'
 
-import type { Model, Collection } from '..'
-import { type RawRecord, type DirtyRaw, sanitizedRaw } from '../RawRecord'
+import type { Model, Collection, Database } from '../..'
+import { type RawRecord, type DirtyRaw, sanitizedRaw } from '../../RawRecord'
 
 // Returns raw record with naive solution to a conflict based on local `_changed` field
 // This is a per-column resolution algorithm. All columns that were changed locally win
@@ -66,4 +66,11 @@ export function prepareMarkAsSynced<T: Model>(record: T): T {
   return record.prepareUpdate(() => {
     replaceRaw(record, newRaw)
   })
+}
+
+export function ensureActionsEnabled(database: Database): void {
+  invariant(
+    database._actionsEnabled,
+    '[Sync] To use Sync, Actions must be enabled. Pass `{ actionsEnabled: true }` to Database constructor — see docs for more details',
+  )
 }
